requirements.txt
-f fi
altair==4.1.0
appdirs==1.4.4
argon2-cffi==20.1.0
astor==0.8.1
astroid==2.4.2
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
base58==2.1.0
beautifulsoup4==4.9.3
bleach==3.3.0
blinker==1.4
blis==0.7.7
cachetools==4.2.2
catalogue==2.0.7
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
cymem==2.0.6
decorator==5.0.7
defusedxml==0.7.1
distlib==0.3.1
docutils==0.16
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.2.0/en_core_web_lg-3.2.0-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.0.12
Flask==1.1.2
gitdb==4.0.7
GitPython==3.1.14
google-search==1.1.1
google-search-results==2.1.0
idna==2.10
importlib-metadata==4.0.1
ipykernel==5.5.4
ipython==7.23.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.5.4
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
joblib==0.17.0
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kalna==0.0.9
keyring==21.4.0
kiwisolver==1.2.0
langcodes==3.3.0
lazy-object-proxy==1.4.3
lxml==4.6.3
MarkupSafe==1.1.1
matplotlib==3.3.2
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
murmurhash==1.0.6
mysql-connector-python==8.0.27
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.20.2
opencv-python==4.5.1.48
openpyxl==3.0.7
packaging==21.3
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pathy==0.6.1
pickleshare==0.7.5
Pillow==8.2.0
pkginfo==1.5.0.1
preshed==3.0.6
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.16.0
pyarrow==4.0.0
pycparser==2.20
pydantic==1.8.2
pydeck==0.6.2
Pygments==2.9.0
pylint==2.6.0
pymongo==4.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytesseract==0.3.9
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
pywin32-ctypes==0.2.0
pywinpty==1.0.1
pyzmq==22.0.3
readme-renderer==26.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.4.0
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.0
secure-smtplib==0.1.1
Send2Trash==1.5.0
six==1.16.0
sklearn==0.0
smart-open==5.2.1
smmap==4.0.0
soupsieve==2.2.1
spacy==3.2.4
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
streamlit==0.81.1
terminado==0.9.4
testpath==0.4.4
thinc==8.0.15
threadpoolctl==2.1.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.50.0
traitlets==5.0.5
twine==3.2.0
typer==0.4.1
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.4
validators==0.18.2
virtualenv==20.4.3
wasabi==0.9.1
watchdog==2.1.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.4.1
